Now support more than 2 params (err, data)
The default async callback handler is 'callback(err, data)', but there are some situations where the callback will get invoked with more than 1 param, such as 'callback(err, path, fd, cleanupCb)'. This is the case in the 'Tmp' package where the callback when creating a temp file is: 'tmp.file([params], function(err, path, fd, cleanupCallback)'. In such a case, we were not able to retrieve the 2 additional values. But with this fix, we can now do so.